apps/api/pyproject.toml
[tool.poetry]
name = "wellchart-new"
version = "0.1.0"
description = ""
authors = ["zacharypfiz <ztpfizcode@gmail.com>"]
readme = "../../README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pymupdf = "^1.24.2"
pandas = "^2.2.2"
fastapi = "^0.111.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
